add cvars + restructure

+++ cli/src/util/debug_draw_manager.h
@@ -129,7 +129,7 @@ struct DebugDrawListT : public GenSimpleHierarchy<TList> {
 	typedef DebugDrawListPolicy<TList> Policy;
 	void clear() { Policy::clear(*this); }
 	void pop_back() { Policy::pop_back(*this); }
-	int size() const { return ((typename TList::Head*)this)->size(); }
+	size_t size() const { return ((typename TList::Head*)this)->size(); }
 	bool validate() const { return Policy::validate(*this); }
 };
